+++ react/src/components/supertable.tsx
@@ -3,8 +3,8 @@ import React, { CSSProperties, Fragment, ReactNode, useMemo } from 'react'
 import { RelativeLoader } from '../navigation/loading'
 import { useColors } from '../page_template/colors'
 import { Universe, useUniverse } from '../universe'
-import { HumanReadableName } from '../urban-stats-script/types-values'
 import { assert } from '../utils/defensive'
+import { HumanReadableName } from '../utils/human-readable-name'
 import { Article } from '../utils/protos'
 
 import { CongressionalColumnData, CongressionalRepresentativeEntry } from './congressional-table/model'
