+++ react/src/components/table.tsx
@@ -10,10 +10,9 @@ import { Colors } from '../page_template/color-themes'
 import { colorFromCycle, useColors } from '../page_template/colors'
 import { MobileArticlePointers, rowExpandedKey, useSetting, useSettings } from '../page_template/settings'
 import { Universe, useUniverse } from '../universe'
-import { reifyReact, reifyString } from '../urban-stats-script/derive-human-readable-name'
-import { HumanReadableName } from '../urban-stats-script/types-values'
 import { withButtonRole } from '../utils/a11y'
 import { assert } from '../utils/defensive'
+import { HumanReadableName, reifyReact, reifyString } from '../utils/human-readable-name'
 import { sanitize } from '../utils/paths'
 import { useComparisonHeadStyle, useMobileLayout } from '../utils/responsive'
 import { isAllowedToBeShown } from '../utils/restricted-types'
