+++ react/src/mapper/components/Colorbar.tsx
@@ -5,9 +5,8 @@ import { getUnitDisplay } from '../../components/unit-display'
 import { Colors } from '../../page_template/color-themes'
 import { useColors } from '../../page_template/colors'
 import { ScaleInstance } from '../../urban-stats-script/constants/scale'
-import { reifyReact, reifyString } from '../../urban-stats-script/derive-human-readable-name'
-import { HumanReadableName } from '../../urban-stats-script/types-values'
 import { furthestColor, interpolateColor } from '../../utils/color'
+import { HumanReadableName, reifyReact, reifyString } from '../../utils/human-readable-name'
 import { classifyStatistic, UnitType } from '../../utils/unit'
 import { Keypoints } from '../ramps'
 import { Basemap } from '../settings/utils'
