+++ react/src/mapper/map-generator.tsx
@@ -26,13 +26,14 @@ import { TextBox } from '../urban-stats-script/constants/text-box'
 import { deriveMapLabel } from '../urban-stats-script/derive-human-readable-name'
 import { EditorError } from '../urban-stats-script/editor-utils'
 import { noLocation } from '../urban-stats-script/location'
-import { HumanReadableName, TypeEnvironment, USSOpaqueValue } from '../urban-stats-script/types-values'
+import { TypeEnvironment, USSOpaqueValue } from '../urban-stats-script/types-values'
 import { AssignmentsResult, executeAsync } from '../urban-stats-script/workerManager'
 import { loadImage } from '../utils/Image'
 import { editIndex, EditSeq } from '../utils/array-edits'
 import { furthestColor, interpolateColor } from '../utils/color'
 import { computeAspectRatioForInsets } from '../utils/coordinates'
 import { makeDebugLogger } from '../utils/debug-logging'
+import { HumanReadableName } from '../utils/human-readable-name'
 import { ConsolidatedShapes, Feature, ICoordinate } from '../utils/protos'
 import { NormalizeProto } from '../utils/types'
 import { useDebouncedResolve } from '../utils/useDebouncedResolve'
