+++ react/src/stat/AddColumnSearchBox.tsx
@@ -4,8 +4,8 @@ import { GenericSearchBox } from '../components/search-generic'
 import { possibilities } from '../mapper/settings/parseExpr'
 import { Selection } from '../mapper/settings/selector-classifier'
 import { addColumn } from '../urban-stats-script/add-column'
-import { reifyReact, reifyString } from '../urban-stats-script/derive-human-readable-name'
-import { HumanReadableName, TypeEnvironment } from '../urban-stats-script/types-values'
+import { TypeEnvironment } from '../urban-stats-script/types-values'
+import { HumanReadableName, reifyReact, reifyString } from '../utils/human-readable-name'
 
 import { Statistic, StatSetter } from './types'
 import { mapUSSFromStat } from './utils'
