+++ react/src/urban-stats-script/constants/constants.ts
@@ -1,6 +1,7 @@
+import { HumanReadableName } from '../../utils/human-readable-name'
 import { Context } from '../context'
 import { hre } from '../human-readable-template'
-import { renderType, USSRawValue, USSValue, DocumentationTable, createConstantExpression, HumanReadableName } from '../types-values'
+import { renderType, USSRawValue, USSValue, DocumentationTable, createConstantExpression } from '../types-values'
 
 import { osmBasemap, noBasemap } from './basemap'
 import { hsv, renderColor, rgb, colorConstants } from './color'
