+++ react/src/urban-stats-script/constants/table.ts
@@ -1,8 +1,9 @@
 import { assert } from '../../utils/defensive'
+import { HumanReadableName } from '../../utils/human-readable-name'
 import { UnitType } from '../../utils/unit'
 import { Context } from '../context'
 import { noLocation } from '../location'
-import { USSType, USSValue, USSRawValue, OriginalFunctionArgs, NamedFunctionArgumentWithDocumentation, createConstantExpression, HumanReadableName } from '../types-values'
+import { USSType, USSValue, USSRawValue, OriginalFunctionArgs, NamedFunctionArgumentWithDocumentation, createConstantExpression } from '../types-values'
 
 export interface TableColumn {
     name: HumanReadableName
