+++ react/src/urban-stats-script/derive-human-readable-name.ts
@@ -1,16 +1,13 @@
-import React, { ReactNode } from 'react'
-
 import { MapUSS, mapUssParser } from '../mapper/settings/map-uss'
 import { assert } from '../utils/defensive'
+import { HumanReadableElement, HumanReadableName } from '../utils/human-readable-name'
 import { separateNumber } from '../utils/text'
 
 import { UrbanStatsASTExpression, UrbanStatsASTStatement } from './ast'
 import * as l from './literal-parser'
 import { noLocation } from './location'
 import { expressionOperatorMap } from './operators'
-import { HumanReadableName, TypeEnvironment } from './types-values'
-
-export type HumanReadableElement = { type: 'atom', value: string } | { type: 'where' | 'superscript' | 'subscript' | 'parens', value: HumanReadableElement[] }
+import { TypeEnvironment } from './types-values'
 
 function humanReadableElements(ast: UrbanStatsASTExpression | UrbanStatsASTStatement, typeEnvironment: TypeEnvironment): HumanReadableElement[] | undefined {
     switch (ast.type) {
@@ -178,53 +175,6 @@ function humanReadableElements(ast: UrbanStatsASTExpression | UrbanStatsASTState
         case 'parseError':
             return undefined
     }
-}
-
-export function reifyReact(elements: HumanReadableElement[] | string): ReactNode {
-    if (typeof elements === 'string') return elements
-    return elements.map((element, index) => {
-        switch (element.type) {
-            case 'atom':
-                return element.value
-            case 'subscript':
-                return <sub key={index}>{reifyReact(element.value)}</sub>
-            case 'superscript':
-                return <sup key={index}>{reifyReact(element.value)}</sup>
-            case 'where':
-                return (
-                    <React.Fragment key={index}>
-                        {' where '}
-                        {reifyReact(element.value)}
-                    </React.Fragment>
-                )
-            case 'parens':
-                return (
-                    <React.Fragment key={index}>
-                        (
-                        {reifyReact(element.value)}
-                        )
-                    </React.Fragment>
-                )
-        }
-    })
-}
-
-export function reifyString(elements: HumanReadableElement[] | string): string {
-    if (typeof elements === 'string') return elements
-    return elements.map((element) => {
-        switch (element.type) {
-            case 'atom':
-                return element.value
-            case 'subscript':
-                return `_{${reifyString(element.value)}}`
-            case 'superscript':
-                return `^{${reifyString(element.value)}}`
-            case 'where':
-                return ` where ${reifyString(element.value)}`
-            case 'parens':
-                return `(${reifyString(element.value)})`
-        }
-    }).join('')
 }
 
 export function deriveMapLabel(uss: MapUSS, typeEnvironment: TypeEnvironment): HumanReadableName | undefined {
